+++ vue3-onsenui/src/components/VOnsRange.vue
@@ -7,12 +7,12 @@
 <script>
   /* This file was generated automatically by 'generate-components' task in bindings/vue/gulpfile.babel.js */
   import 'onsenui/esm/elements/ons-range.js';
-  import { deriveEvents, modelInput, modifier } from '../mixins/index.js';
+  import { deriveEvents, modelInputNumber, modifier } from '../mixins/index.js';
 
   const name = 'v-ons-range';
 
   export default {
     name,
-    mixins: [deriveEvents(name), modelInput, modifier]
+    mixins: [deriveEvents(name), modelInputNumber, modifier]
   };
 </script>