+++ vue3-onsenui/src/mixins/index.js
@@ -1,3 +1,3 @@
 export { deriveDBB, deriveEvents, unrecognizedListeners } from './derive';
 export { hidable, hasOptions, selfProvider, dialogCancel, portal, modifier } from './common';
-export { modelInput, modelCheckbox, modelRadio } from './model';
+export { modelInput, modelInputNumber, modelCheckbox, modelRadio } from './model';